Commiting intermittent updates. Will do a pull request for task # 9 for training programs in the morning
