+++ BangazonWorkforce/Controllers/PaymentTypeController.cs
@@ -3,12 +3,12 @@ using System.Collections.Generic;
 using System.Data.SqlClient;
 using System.Linq;
 using System.Threading.Tasks;
-using BangazonAPI.Models;
+using BangazonWorkforce.Models;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 
-namespace BangazonAPI.Controllers
+namespace BangazonWorkforce.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
