+++ BangazonWorkforce/Models/Customer.cs
@@ -5,7 +5,7 @@
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 
-namespace BangazonAPI.Models
+namespace BangazonWorkforce.Models
 {
     public class Customer
     {
