+++ BangazonWorkforce/appsettings.Development.json
@@ -5,5 +5,9 @@
       "System": "Information",
       "Microsoft": "Information"
     }
+  },
+  "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "DefaultConnection": "Server=localhost\\SQLEXPRESS;Database=BangazonWorkforce;Trusted_Connection=True;"
   }
 }
