+++ BangazonWorkforce/appsettings.json
@@ -4,8 +4,5 @@
       "Default": "Warning"
     }
   },
-  "AllowedHosts": "*",
-  "ConnectionStrings": {
-    "DefaultConnection": "Server=localhost\\SQLEXPRESS;Database=BangazonWorkforce;Trusted_Connection=True;"
+  "AllowedHosts": "*"
   }
-}
